4.3 create Tooltip component and small adjustments
