+++ src/components/Marker.js
@@ -1,6 +1,6 @@
 import React from "react";
-import marker from '../push_pin-black-18dp.svg';
+import marker from '../assets/images/push_pin-black-18dp.svg';
 
-const MarkerComponent = () => <img className="map__marker" src={marker} alt="Marker" />;
+const MarkerComponent = () => <img style={{transform: 'translate(-50%, -100%)'}} className="map__marker" src={marker} alt="Marker" />;
 
 export default MarkerComponent;
